+++ src/index.ts
@@ -50,4 +50,4 @@ export {
   PAYKU_PAYMENT_METHODS,
   PAYKU_VES_GATEWAYS,
 } from "./constants/payku.constants";
-export { buildPaymentRedirectUrl } from "./utils/payku.utils";
+export { buildPaymentRedirectUrl, buildMallMerchant } from "./utils/payku.utils";